+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "SwiftTestFileGen" extension will be documented in this file.
 
+## [Unreleased]
+
+- Disallowing special characters in patterns for `swiftTestFileGen.gotoTestFile.heuristicFilenamePattern`.
+
 ## [0.3.0]
 
 - Fixing system, binary, and plugin targets not being recognized and resulting in Package.swift parse errors.